requirements.txt
streamlit==1.29.0
tensorflow==2.13.0
pillow==10.0.0
numpy==1.26.0
